move triggerDetection in other file
